+++ config/child-theme-settings.php
@@ -1,0 +1,37 @@
+<?php
+/**
+ * Genesis Advanced
+ *
+ * This file adds the child theme settings
+ *
+ * @package Genesis Advanced
+ * @author  NicBeltramelli
+ * @license GPL-2.0-or-later
+ * @link    https://github.com/NicBeltramelli/genesis-advanced.git
+ */
+
+if ( ! defined( 'ABSPATH' ) ) {
+	exit; // Exit if accessed directly.
+}
+
+return array(
+	GENESIS_SETTINGS_FIELD => array(
+		'blog_cat_num'              => 6,
+		'breadcrumb_home'           => 0,
+		'breadcrumb_front_page'     => 0,
+		'breadcrumb_posts_page'     => 0,
+		'breadcrumb_single'         => 0,
+		'breadcrumb_page'           => 0,
+		'breadcrumb_archive'        => 0,
+		'breadcrumb_404'            => 0,
+		'breadcrumb_attachment'     => 0,
+		'content_archive'           => 'full',
+		'content_archive_limit'     => 0,
+		'content_archive_thumbnail' => 0,
+		'image_size'                => '',
+		'image_alignment'           => 'alignleft',
+		'posts_nav'                 => 'numeric',
+		'site_layout'               => 'content-sidebar',
+	),
+	'posts_per_page'       => 6,
+);